+++ CHANGES.rst
@@ -10,6 +10,7 @@ Changes
 - Geometry based filtering of feature tables on webmaps.
 - Passing a zoom level to webmaps via ``zoom`` attribute.
 - Ability to refresh feature tables from UI.
+- Support for OpenID Connect UserInfo endpoint.